Update files (#84)
* fix shellcheck SC2115

* update readme text
remove plugin creation step

* add codefactor
add direct link for check_mk agent v2.3.0pXX
changed requirements text

+++ source/compile_docker.sh
@@ -34,4 +34,4 @@ md5sum check_mk_agent-"$(date +'%Y.%m.%d')".$VERSION.tgz > check_mk_agent-"$(dat
 
 # Move to packages folder and cleanup
 mv check_mk_agent-"$(date +'%Y.%m.%d')".$VERSION.tgz check_mk_agent-"$(date +'%Y.%m.%d')".$VERSION.tgz.md5 /build/packages
-rm -rf ${DATA_DIR}/*
+rm -rf ${DATA_DIR:?}/*
